backend folder structured, mongodb connected, uri should be add environment and frontend api path changed with updated versionans.
