+++ backend/src/config/mongo.js
@@ -1,0 +1,17 @@
+const mongoose = require("mongoose");
+const { options } = require("../routes");
+require("dotenv").config();
+
+const connectDB = async () => {
+  try {
+    await mongoose.connect(
+      process.env.MONGODB_URI || "mongodb://127.0.0.1:27017/"
+    );
+    console.log("MongoDB bağlantısı başarılı");
+  } catch (error) {
+    console.error("MongoDB bağlantı hatası:", error);
+    process.exit(1);
+  }
+};
+
+module.exports = connectDB;
