+++ backend/src/routes/auth.routes.js
@@ -1,0 +1,8 @@
+const express = require("express");
+const router = express.Router();
+const authController = require("../controllers/auth.controller");
+
+router.post("/login", authController.login);
+router.post("/logout", authController.logout);
+
+module.exports = router;
