+++ backend/src/routes/user.routes.js
@@ -1,0 +1,7 @@
+const express = require("express");
+const router = express.Router();
+const userController = require("../controllers/user.controller");
+
+router.get("/", userController.getUser);
+
+module.exports = router;
